requirements.txt
Django==3.1.3
django-filter==2.4.0
django-graphql-jwt==0.3.1
PyJWT==1.7.1
graphene-django==2.13.0
django-cors-headers==3.5.0
django-health-check
python-decouple
setuptools
psycopg2-binary==2.9.9
